add user mapping and edit functionality

+++ src/main/java/com/example/url_shortner/repository/UrlRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UrlRepository extends JpaRepository<UrlModel,Long> {
+public interface UrlRepository extends JpaRepository<UrlModel,Integer> {
 
     Optional<UrlModel> findByShortCode(String shortCode);
